// src/admin/maestro-establecimiento/maestro-establecimiento.interface.ts
export interface MaestroEstablecimiento {

    Id_Establecimiento: number;
    Nombre_Establecimiento: string;
    Ubigueo_Establecimiento: number;
    Codigo_Disa: number;
    Disa: string;
    Codigo_Red: string;
    Red: string;
    Codigo_MicroRed: string;
    MicroRed: string;
    Codigo_Unico: string;
    Codigo_Sector: number;
    Descripcion_Sector: string;
    Departamento: string;
    Provincia: string;
    Distrito: string;
}